Fix graphs generation with column mismatch
With a mismatch in columns, gnuplot would previously generate
nonsensical graphs containing multiple columns into a single column.

This patch fixes this issue by filtering the latest 10 bencharks to
ignore all files which have less columns than the benchmark with the
most columns.

Theoretically it would still be possible to create issues by adding and
removing a benchmark in under 10 master commits, which would cause the
results to appear in an unrelated column. However since this is unlikely
and unnecessary at the moment, it's not considered for this PR.